@@ -5,7 +5,7 @@ homepage: https://github.com/terrabythia/dart_chromecast
 publish_to: none
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
   flutter: '>=1.20.0'
 
 dependencies:
