@@ -8,15 +8,15 @@ import '../proto/cast_channel.pb.dart';
 abstract class CastChannel {
   static int _requestId = 1;
 
-  final Socket _socket;
-  String _sourceId;
-  String _destinationId;
-  String _namespace;
+  final Socket? _socket;
+  String? _sourceId;
+  String? _destinationId;
+  String? _namespace;
 
   CastChannel(this._socket, this._sourceId, this._destinationId, this._namespace);
 
-  CastChannel.createWithSocket(Socket socket,
-      {String sourceId, String destinationId, String namespace})
+  CastChannel.createWithSocket(Socket? socket,
+      {String? sourceId, String? destinationId, String? namespace})
       : _socket = socket,
         _sourceId = sourceId,
         _destinationId = destinationId,
@@ -27,9 +27,9 @@ abstract class CastChannel {
 
     CastMessage castMessage = CastMessage();
     castMessage.protocolVersion = CastMessage_ProtocolVersion.CASTV2_1_0;
-    castMessage.sourceId = _sourceId;
-    castMessage.destinationId = _destinationId;
-    castMessage.namespace = _namespace;
+    castMessage.sourceId = _sourceId!;
+    castMessage.destinationId = _destinationId!;
+    castMessage.namespace = _namespace!;
     castMessage.payloadType = CastMessage_PayloadType.STRING;
     castMessage.payloadUtf8 = jsonEncode(payload);
 
@@ -45,7 +45,7 @@ abstract class CastChannel {
       print('PING');
     }
 
-    _socket.add(fullData);
+    _socket!.add(fullData);
 
     _requestId++;
   }
