@@ -1,11 +1,11 @@
 class CastMedia {
 
-  final String contentId;
-  String title;
+  final String? contentId;
+  String? title;
   bool autoPlay = true;
   double position;
   String contentType;
-  List<String> images;
+  List<String>? images;
 
   CastMedia({
     this.contentId,
