@@ -22,23 +22,23 @@ class CastSender extends Object {
   final Logger log = const Logger('CastSender');
   final CastDevice device;
 
-  SecureSocket _socket;
+  SecureSocket? _socket;
 
-  ConnectionChannel _connectionChannel;
-  HeartbeatChannel _heartbeatChannel;
-  ReceiverChannel _receiverChannel;
-  MediaChannel _mediaChannel;
+  ConnectionChannel? _connectionChannel;
+  HeartbeatChannel? _heartbeatChannel;
+  ReceiverChannel? _receiverChannel;
+  MediaChannel? _mediaChannel;
 
-  bool connectionDidClose;
+  late bool connectionDidClose;
 
 //  Timer _heartbeatTimer;
-  Timer _mediaCurrentTimeTimer;
+  Timer? _mediaCurrentTimeTimer;
 
-  CastSession _castSession;
-  StreamController<CastSession> castSessionController;
-  StreamController<CastMediaStatus> castMediaStatusController;
-  List<CastMedia> _contentQueue;
-  CastMedia _currentCastMedia;
+  CastSession? _castSession;
+  late StreamController<CastSession?> castSessionController;
+  late StreamController<CastMediaStatus?> castMediaStatusController;
+  late List<CastMedia> _contentQueue;
+  CastMedia? _currentCastMedia;
 
   CastSender(this.device) {
     // TODO: _airplay._tcp
@@ -62,7 +62,7 @@ class CastSender extends Object {
       return false;
     }
 
-    _connectionChannel.sendMessage({'type': 'CONNECT'});
+    _connectionChannel!.sendMessage({'type': 'CONNECT'});
 
     // start heartbeat
     _heartbeatTick();
@@ -74,7 +74,7 @@ class CastSender extends Object {
     return true;
   }
 
-  Future<bool> reconnect({String sourceId, String destinationId}) async {
+  Future<bool> reconnect({String? sourceId, String? destinationId}) async {
     _castSession = CastSession(sourceId: sourceId, destinationId: destinationId);
     bool connected = await connect();
     if (!connected) {
@@ -83,7 +83,7 @@ class CastSender extends Object {
 
     _mediaChannel =
         MediaChannel.create(socket: _socket, sourceId: sourceId, destinationId: destinationId);
-    _mediaChannel.sendMessage({'type': 'GET_STATUS'});
+    _mediaChannel!.sendMessage({'type': 'GET_STATUS'});
 
     // now wait for the media to actually get a status?
     bool didReconnect = await _waitForMediaStatus();
@@ -99,7 +99,7 @@ class CastSender extends Object {
       }
 
       try {
-        castMediaStatusController.add(_castSession.castMediaStatus);
+        castMediaStatusController.add(_castSession!.castMediaStatus);
       } catch (e) {
         log.severe(
             "Could not add the CastMediaStatus to the CastSession Stream Controller: events will not be triggered");
@@ -112,9 +112,9 @@ class CastSender extends Object {
 
   Future<bool> disconnect() async {
     if (null != _connectionChannel && null != _castSession?.castMediaStatus) {
-      _connectionChannel.sendMessage({
+      _connectionChannel!.sendMessage({
         'type': 'CLOSE',
-        'sessionId': _castSession.castMediaStatus.sessionId,
+        'sessionId': _castSession!.castMediaStatus!.sessionId,
       });
     }
     _socket?.destroy();
@@ -123,9 +123,9 @@ class CastSender extends Object {
     return true;
   }
 
-  void launch([String appId]) {
+  void launch([String? appId]) {
     if (null != _receiverChannel) {
-      _receiverChannel.sendMessage({
+      _receiverChannel!.sendMessage({
         'type': 'LAUNCH',
         'appId': appId ?? 'CC1AD845',
       });
@@ -147,12 +147,12 @@ class CastSender extends Object {
     }
   }
 
-  void _castMediaAction(String type, [Map<String, dynamic> params]) {
+  void _castMediaAction(String type, [Map<String, dynamic>? params]) {
     params ??= {};
     if (null != _mediaChannel && null != _castSession?.castMediaStatus) {
-      _mediaChannel.sendMessage(params
+      _mediaChannel!.sendMessage(params
         ..addAll({
-          'mediaSessionId': _castSession.castMediaStatus.sessionId,
+          'mediaSessionId': _castSession!.castMediaStatus!.sessionId,
           'type': type,
         }));
     }
@@ -170,7 +170,6 @@ class CastSender extends Object {
 
   void togglePause() {
     log.info("TOGGLE_PAUSE");
-    log.info(_castSession?.castMediaStatus.toString());
     if (true == _castSession?.castMediaStatus?.isPlaying) {
       pause();
     } else if (true == _castSession?.castMediaStatus?.isPaused) {
@@ -192,27 +191,27 @@ class CastSender extends Object {
     _castMediaAction('VOLUME', map);
   }
 
-  CastSession get castSession => _castSession;
+  CastSession? get castSession => _castSession;
 
   // private
 
-  Future<SecureSocket> _createSocket() async {
+  Future<SecureSocket?> _createSocket() async {
     if (null == _socket) {
       try {
         log.fine('Connecting to ${device.host}:${device.port}');
 
-        _socket = await SecureSocket.connect(device.host, device.port,
+        _socket = await SecureSocket.connect(device.host, device.port!,
             onBadCertificate: (X509Certificate certificate) => true,
             timeout: Duration(seconds: 10));
 
         _connectionChannel = ConnectionChannel.create(_socket,
-            sourceId: _castSession.sourceId, destinationId: _castSession.destinationId);
+            sourceId: _castSession!.sourceId, destinationId: _castSession!.destinationId);
         _heartbeatChannel = HeartbeatChannel.create(_socket,
-            sourceId: _castSession.sourceId, destinationId: _castSession.destinationId);
+            sourceId: _castSession!.sourceId, destinationId: _castSession!.destinationId);
         _receiverChannel = ReceiverChannel.create(_socket,
-            sourceId: _castSession.sourceId, destinationId: _castSession.destinationId);
+            sourceId: _castSession!.sourceId, destinationId: _castSession!.destinationId);
 
-        _socket.listen(_onSocketData, onDone: _dispose);
+        _socket!.listen(_onSocketData, onDone: _dispose);
       } catch (e) {
         log.fine(e.toString());
         return null;
@@ -225,21 +224,16 @@ class CastSender extends Object {
     List<int> slice = event.getRange(4, event.length).toList();
 
     CastMessage message = CastMessage.fromBuffer(slice);
-    if (null != message) {
-      // handle the message
-      if (null != message.payloadUtf8) {
-        Map<String, dynamic> payloadMap = jsonDecode(message.payloadUtf8);
-        log.fine(payloadMap['type']);
-        if ('CLOSE' == payloadMap['type']) {
-          _dispose();
-          connectionDidClose = true;
-        }
-        if ('RECEIVER_STATUS' == payloadMap['type']) {
-          _handleReceiverStatus(payloadMap);
-        } else if ('MEDIA_STATUS' == payloadMap['type']) {
-          _handleMediaStatus(payloadMap);
-        }
-      }
+    Map<String, dynamic> payloadMap = jsonDecode(message.payloadUtf8);
+    log.fine(payloadMap['type']);
+    if ('CLOSE' == payloadMap['type']) {
+      _dispose();
+      connectionDidClose = true;
+    }
+    if ('RECEIVER_STATUS' == payloadMap['type']) {
+      _handleReceiverStatus(payloadMap);
+    } else if ('MEDIA_STATUS' == payloadMap['type']) {
+      _handleMediaStatus(payloadMap);
     }
   }
 
@@ -247,17 +241,17 @@ class CastSender extends Object {
     log.fine(payload.toString());
     if (null == _mediaChannel && true == payload['status']?.containsKey('applications')) {
       // re-create the channel with the transportId the chromecast just sent us
-      if (false == _castSession?.isConnected) {
-        _castSession = _castSession
+      if (false == _castSession!.isConnected) {
+        _castSession = _castSession!
           ..mergeWithChromeCastSessionMap(payload['status']['applications'][0]);
         _connectionChannel = ConnectionChannel.create(_socket,
-            sourceId: _castSession.sourceId, destinationId: _castSession.destinationId);
-        _connectionChannel.sendMessage({'type': 'CONNECT'});
+            sourceId: _castSession!.sourceId, destinationId: _castSession!.destinationId);
+        _connectionChannel!.sendMessage({'type': 'CONNECT'});
         _mediaChannel = MediaChannel.create(
             socket: _socket,
-            sourceId: _castSession.sourceId,
-            destinationId: _castSession.destinationId);
-        _mediaChannel.sendMessage({'type': 'GET_STATUS'});
+            sourceId: _castSession!.sourceId,
+            destinationId: _castSession!.destinationId);
+        _mediaChannel!.sendMessage({'type': 'GET_STATUS'});
 
         try {
           castSessionController.add(_castSession);
@@ -271,41 +265,41 @@ class CastSender extends Object {
   }
 
   Future<bool> _waitForMediaStatus() async {
-    while (false == _castSession.isConnected) {
+    while (false == _castSession!.isConnected) {
       await Future.delayed(Duration(milliseconds: 100));
       if (connectionDidClose) return false;
     }
-    return _castSession.isConnected;
+    return _castSession!.isConnected;
   }
 
   void _handleMediaStatus(Map payload) {
     log.fine('Handle media status: ' + payload.toString());
 
     if (null != payload['status']) {
-      if (!_castSession.isConnected) {
-        _castSession.isConnected = true;
+      if (!_castSession!.isConnected) {
+        _castSession!.isConnected = true;
         _handleContentQueue();
       }
 
       if (payload['status'].length > 0) {
-        _castSession.castMediaStatus =
+        _castSession!.castMediaStatus =
             CastMediaStatus.fromChromeCastMediaStatus(payload['status'][0]);
 
-        log.fine('Media status ${_castSession.castMediaStatus.toString()}');
+        log.fine('Media status ${_castSession!.castMediaStatus.toString()}');
 
-        if (_castSession.castMediaStatus.isFinished) {
+        if (_castSession!.castMediaStatus!.isFinished) {
           _handleContentQueue();
         }
 
-        if (_castSession.castMediaStatus.isPlaying) {
+        if (_castSession!.castMediaStatus!.isPlaying) {
           _mediaCurrentTimeTimer = Timer(Duration(seconds: 1), _getMediaCurrentTime);
-        } else if (_castSession.castMediaStatus.isPaused && null != _mediaCurrentTimeTimer) {
-          _mediaCurrentTimeTimer.cancel();
+        } else if (_castSession!.castMediaStatus!.isPaused && null != _mediaCurrentTimeTimer) {
+          _mediaCurrentTimeTimer!.cancel();
           _mediaCurrentTimeTimer = null;
         }
 
         try {
-          castMediaStatusController.add(_castSession.castMediaStatus);
+          castMediaStatusController.add(_castSession!.castMediaStatus);
         } catch (e) {
           log.severe(
               "Could not add the CastMediaStatus to the CastSession Stream Controller: events will not be triggered");
@@ -327,8 +321,8 @@ class CastSender extends Object {
     if (null == _mediaChannel || _contentQueue.isEmpty) {
       return;
     }
-    if (null != _castSession.castMediaStatus &&
-        !_castSession.castMediaStatus.isFinished &&
+    if (null != _castSession!.castMediaStatus &&
+        !_castSession!.castMediaStatus!.isFinished &&
         !forceNext) {
       // don't handle the next in the content queue, because we only want
       // to play the 'next' content if it's not already playing.
@@ -337,13 +331,13 @@ class CastSender extends Object {
     _currentCastMedia = _contentQueue.elementAt(0);
     if (null != _currentCastMedia) {
       _contentQueue = _contentQueue.getRange(1, _contentQueue.length).toList();
-      _mediaChannel.sendMessage(_currentCastMedia.toChromeCastMap());
+      _mediaChannel!.sendMessage(_currentCastMedia!.toChromeCastMap());
     }
   }
 
   void _getMediaCurrentTime() {
     if (null != _mediaChannel && true == _castSession?.castMediaStatus?.isPlaying) {
-      _mediaChannel.sendMessage({
+      _mediaChannel!.sendMessage({
         'type': 'GET_STATUS',
       });
     }
@@ -351,7 +345,7 @@ class CastSender extends Object {
 
   void _heartbeatTick() {
     if (null != _heartbeatChannel) {
-      _heartbeatChannel.sendMessage({'type': 'PING'});
+      _heartbeatChannel!.sendMessage({'type': 'PING'});
 
 //      _heartbeatTimer = Timer(Duration(seconds: 5), _heartbeatTick);
       Timer(Duration(seconds: 5), _heartbeatTick);
